Maidenhead beacon test updates.
